+++ esp-hal-common/src/ledc/channel.rs
@@ -136,10 +136,10 @@ pub trait ChannelHW<O: OutputPin> {
 
 /// Channel struct
 pub struct Channel<'a, S: TimerSpeed, O: OutputPin> {
-    pub(crate) ledc: &'a RegisterBlock,
-    pub(crate) timer: Option<&'a dyn TimerIFace<S>>,
-    pub(crate) number: Number,
-    pub(crate) output_pin: PeripheralRef<'a, O>,
+    pub ledc: &'a RegisterBlock,
+    pub timer: Option<&'a dyn TimerIFace<S>>,
+    pub number: Number,
+    pub output_pin: PeripheralRef<'a, O>,
 }
 
 impl<'a, S: TimerSpeed, O: OutputPin> Channel<'a, S, O> {