+++ CHANGELOG.md
@@ -32,11 +32,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add a `defmt` feature to the `esp-hal-smartled` package (#846)
 - Support 16MB octal PS-RAM for ESP32-S3 (#858)
 - RISCV TRACE Encoder driver for ESP32-C6 / ESP32-H2 (#864)
+- Add Buzzer driver into `esp-hal-buzzer` (#814)
 
 ### Changed
 
 - Bumped MSRV to 1.67 (#798)
 - Optimised multi-core critical section implementation (#797)
+- Move `esp-hal-smartled` into `extras/` (#814)
 
 ### Fixed